+++ CHANGELOG.md
@@ -3,7 +3,22 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Changelog](http://keepachangelog.com/).
 
+
 ## Unreleased
+---
+
+### New
+
+- [#2754](https://gitlab.com/meltano/meltano/-/issues/2754) New `meltano state` command to manage job states.
+
+### Changes
+
+### Fixes
+
+### Breaks
+
+
+## 1.101.0 - (2022-04-21)
 
 ---
 
@@ -11,7 +26,6 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 - [#3377](https://gitlab.com/meltano/meltano/-/issues/3377) Add active environment as variable `MELTANO_ENVIRONMENT` in Plugin execution environment.
 
-### Changes
 
 ### Fixes
 
@@ -20,7 +34,6 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 - [#3390](https://gitlab.com/meltano/meltano/-/issues/3390) Fixes bug where `meltano init` failed to create prerequisite directories.
 - [#3391](https://gitlab.com/meltano/meltano/-/issues/3391) Fixes bug where `meltano run` would invoke the wrong plugin command when the same plugin was used multiple times.
 
-### Breaks
 
 ## 1.100.0 - (2022-04-14)